Fix bug when specifying include and fields.
Previously, the code was applying the *last* `fields` value
specified to *all* types of resource; this meant that other
resources would have their attributes stripped altogether.
This fixes that by keeping a map of fields for each model,
so it can apply the correct list of fields to the resource.