Disallow use of `var` in favor of `let` or `const`
In the previous commit I fixed a bug which appears to have been missed
because of the unexpected scoping rules of `var`:
the `field` variable was declared inside the for loop,
but because of var hoisting it was then used *outside* of the loop
in an inappropriate context.
With the addition of this rule such a mistake is no longer possible.

This appears to have been an existing but informal policy:
other than the aforementioned bug, var was used in only two other places,
both of which I have fixed.